simpler PR comments
