+++ extension/chrome/elements/compose-modules/compose-my-pubkey-module.ts
@@ -8,7 +8,7 @@ import { Ui } from '../../../js/common/browser/ui.js';
 import { ViewModule } from '../../../js/common/view-module.js';
 import { ComposeView } from '../compose.js';
 import { Str } from '../../../js/common/core/common.js';
-import { KeyStoreUtil } from '../../../js/common/platform/store/key-store.js';
+import { KeyStoreUtil } from "../../../js/common/core/crypto/key-store-util";
 
 export class ComposeMyPubkeyModule extends ViewModule<ComposeView> {
