+++ extension/js/common/assert.ts
@@ -41,19 +41,18 @@ export class Assert {
       const parsedKeys = await Promise.all(kis.map(ki => KeyUtil.parse(ki.private)));
       const { setup_done } = await AcctStore.get(acctEmail, ['setup_done']);
       if (setup_done && kis.length) {
-        for (const key of parsedKeys) {
-          if (!key.fullyEncrypted) {
-            if (window.location.pathname === '/chrome/settings/index.htm') {
-              await Settings.renderSubPage(acctEmail, tabId!, '/chrome/settings/modules/change_passphrase.htm');
-            } else {
-              const msg = `Protect your key with a pass phrase to finish setup.`;
-              const r = await Ui.renderOverlayPromptAwaitUserChoice({ finishSetup: {}, later: { color: 'gray' } }, msg, undefined,
-                Lang.general.contactIfNeedAssistance(await isFesUsed(acctEmail)));
-              if (r === 'finish_setup') {
-                await Browser.openSettingsPage('index.htm', acctEmail);
-              }
+        const key = parsedKeys.find(k => !k.fullyEncrypted);
+        if (key) {
+          // can fix one key at a time. When they reload, it will complain about another key
+          if (window.location.pathname === '/chrome/settings/index.htm') {
+            await Settings.renderSubPage(acctEmail, tabId!, '/chrome/settings/modules/change_passphrase.htm');
+          } else {
+            const msg = `Protect your key with a pass phrase to finish setup.`;
+            const r = await Ui.renderOverlayPromptAwaitUserChoice({ finishSetup: {}, later: { color: 'gray' } }, msg, undefined,
+              Lang.general.contactIfNeedAssistance(await isFesUsed(acctEmail)));
+            if (r === 'finish_setup') {
+              await Browser.openSettingsPage('index.htm', acctEmail);
             }
-            return; // can fix one key at a time. When they reload, it will complain about another key
           }
         }
       }
